Add option to specify a data cache path separate from data directory.
Switches the cache to using md5 hashes of a text description instead
of crafted filenames to determine a "cache hit".

Changes the default location of these files to be an "index-cache"
directory inside the data root. Should leave the data directories a
bit cleaner, especially with these filenames being a bit "uglier".

For GPT the code will first look in this default location before
building a new index and caching it the specified data cache path (or
this default if none is given).

For Blendable dataset it will only look for and save the indices if a
data cache path is provided, otherwise it will just rebuild every
time.