Do not tie the output layer with the word embeddings unless specified.
This adds an argument share_word_embeddings_and_output_weights to
GPTModel. It also reworks out word embeddings and output weights are
shared in that case.

An "output_layer" is always created. If it is to share weights with
the word embeddings (and are in the same pipeline rank), then the
weights of the output_layer are not allocated (this is a new option to
ColumnParallelLinear) and the word embedding weights are instead
passed to the output_layer's forward method. If the weights are not
shared, or they are on different pipeline ranks, then the output_layer
allocates its own weights as normal, and those weight are synced with
the first stage's word embedding weights as needed.